finnecto_middleware/pyproject.toml
[project]
name = "finnecto-middleware"
version = "0.1.0"
description = "Integration middleware for multi-system invoice processing"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
